Fix one-page paginator

+++ core/paginator.py
@@ -87,27 +87,31 @@ class PaginatorSession:
     def update_disabled_status(self):
         if self.current == self.first_page():
             # disable << button
-            if self._buttons_map["<<"]:
+            if self._buttons_map["<<"] is not None:
                 self._buttons_map["<<"].disabled = True
 
-            self._buttons_map["<"].disabled = True
+            if self._buttons_map["<"] is not None:
+                self._buttons_map["<"].disabled = True
         else:
-            if self._buttons_map["<<"]:
+            if self._buttons_map["<<"] is not None:
                 self._buttons_map["<<"].disabled = False
 
-            self._buttons_map["<"].disabled = False
+            if self._buttons_map["<"] is not None:
+                self._buttons_map["<"].disabled = False
 
         if self.current == self.last_page():
             # disable >> button
             if self._buttons_map[">>"] is not None:
                 self._buttons_map[">>"].disabled = True
 
-            self._buttons_map[">"].disabled = True
+            if self._buttons_map[">"] is not None:
+                self._buttons_map[">"].disabled = True
         else:
             if self._buttons_map[">>"] is not None:
                 self._buttons_map[">>"].disabled = False
 
-            self._buttons_map[">"].disabled = False
+            if self._buttons_map[">"] is not None:
+                self._buttons_map[">"].disabled = False
 
     async def create_base(self, item) -> None:
         """
@@ -151,7 +155,10 @@ class PaginatorSession:
         """
         if not self.running:
             await self.show_page(self.current)
-            await self.view.wait()
+
+            if self.view is not None:
+                await self.view.wait()
+
             await self.close(delete=False)
 
     async def close(
@@ -182,12 +189,13 @@ class PaginatorSession:
 
             self.running = False
 
-            self.view.stop()
-            if delete:
-                await message.delete()
-            else:
-                self.view.clear_items()
-                await message.edit(view=self.view)
+            if self.view is not None:
+                self.view.stop()
+                if delete:
+                    await message.delete()
+                else:
+                    self.view.clear_items()
+                    await message.edit(view=self.view)
 
 
 class PaginatorView(View):
